+++ src/main/java/cspfj/priorityqueues/FibonacciHeap.java
@@ -203,8 +203,8 @@ public final class FibonacciHeap<T extends Identified> extends AbstractQueue<T> 
             map[id] = node;
         }
 
-        final double oldKey = node.key;
-        final double newKey = key.getKey(data);
+        final float oldKey = node.key;
+        final float newKey = key.getKey(data);
         node.key = newKey;
 
         if (node.inQueue) {
@@ -486,7 +486,7 @@ public final class FibonacciHeap<T extends Identified> extends AbstractQueue<T> 
          */
         private int degree;
 
-        private double key;
+        private float key;
 
         private boolean inQueue = false;
